requirements.txt
openai
openai-whisper
torchaudio
playsound
elevenlabs